Ids Integer2Long: Group, Users, Users2

+++ src/main/java/de/wbstraining/lotto/persistence/model/Users.java
@@ -7,6 +7,7 @@ package de.wbstraining.lotto.persistence.model;
 
 import java.io.Serializable;
 import java.util.List;
+
 import javax.persistence.Basic;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -29,146 +30,148 @@ import javax.validation.constraints.Size;
 @Entity
 @Table(name = "users")
 @NamedQueries({
-    @NamedQuery(name = "Users.findAll", query = "SELECT u FROM Users u"),
-    @NamedQuery(name = "Users.findByUserId", query = "SELECT u FROM Users u WHERE u.userId = :userId"),
-    @NamedQuery(name = "Users.findByUsername", query = "SELECT u FROM Users u WHERE u.username = :username"),
-    @NamedQuery(name = "Users.findByFirstName", query = "SELECT u FROM Users u WHERE u.firstName = :firstName"),
-    @NamedQuery(name = "Users.findByMiddleName", query = "SELECT u FROM Users u WHERE u.middleName = :middleName"),
-    @NamedQuery(name = "Users.findByLastName", query = "SELECT u FROM Users u WHERE u.lastName = :lastName"),
-    @NamedQuery(name = "Users.findByPassword", query = "SELECT u FROM Users u WHERE u.password = :password")})
+		@NamedQuery(name = "Users.findAll", query = "SELECT u FROM Users u"),
+		@NamedQuery(name = "Users.findByUserId", query = "SELECT u FROM Users u WHERE u.userId = :userId"),
+		@NamedQuery(name = "Users.findByUsername", query = "SELECT u FROM Users u WHERE u.username = :username"),
+		@NamedQuery(name = "Users.findByFirstName", query = "SELECT u FROM Users u WHERE u.firstName = :firstName"),
+		@NamedQuery(name = "Users.findByMiddleName", query = "SELECT u FROM Users u WHERE u.middleName = :middleName"),
+		@NamedQuery(name = "Users.findByLastName", query = "SELECT u FROM Users u WHERE u.lastName = :lastName"),
+		@NamedQuery(name = "Users.findByPassword", query = "SELECT u FROM Users u WHERE u.password = :password") })
 public class Users implements Serializable {
 
-    private static final long serialVersionUID = 1L;
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Basic(optional = false)
-    @Column(name = "user_id")
-    private Integer userId;
-    @Basic(optional = false)
-    @NotNull
-    @Size(min = 1, max = 10)
-    @Column(name = "username")
-    private String username;
-    @Size(max = 20)
-    @Column(name = "first_name")
-    private String firstName;
-    @Size(max = 20)
-    @Column(name = "middle_name")
-    private String middleName;
-    @Size(max = 20)
-    @Column(name = "last_name")
-    private String lastName;
-    @Basic(optional = false)
-    @NotNull
-    @Size(min = 1, max = 64)
-    @Column(name = "password")
-    private String password;
-    @ManyToMany(mappedBy = "usersList")
-    private List<Groups> groupsList;
-    @JoinColumn(name = "kundeid", referencedColumnName = "kundeid")
-    @OneToOne
-    private Kunde kunde;
+	private static final long serialVersionUID = 1L;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Basic(optional = false)
+	@Column(name = "user_id")
+	private Long userId;
+	@Basic(optional = false)
+	@NotNull
+	@Size(min = 1, max = 10)
+	@Column(name = "username")
+	private String username;
+	@Size(max = 20)
+	@Column(name = "first_name")
+	private String firstName;
+	@Size(max = 20)
+	@Column(name = "middle_name")
+	private String middleName;
+	@Size(max = 20)
+	@Column(name = "last_name")
+	private String lastName;
+	@Basic(optional = false)
+	@NotNull
+	@Size(min = 1, max = 64)
+	@Column(name = "password")
+	private String password;
+	@ManyToMany(mappedBy = "usersList")
+	private List<Groups> groupsList;
+	@JoinColumn(name = "kundeid", referencedColumnName = "kundeid")
+	@OneToOne
+	private Kunde kunde;
 
-    public Users() {
-    }
+	public Users() {
+	}
 
-    public Users(Integer userId) {
-        this.userId = userId;
-    }
+	public Users(Long userId) {
+		this.userId = userId;
+	}
 
-    public Users(Integer userId, String username, String password) {
-        this.userId = userId;
-        this.username = username;
-        this.password = password;
-    }
+	public Users(Long userId, String username, String password) {
+		this.userId = userId;
+		this.username = username;
+		this.password = password;
+	}
 
-    public Integer getUserId() {
-        return userId;
-    }
+	public Long getUserId() {
+		return userId;
+	}
 
-    public void setUserId(Integer userId) {
-        this.userId = userId;
-    }
+	public void setUserId(Long userId) {
+		this.userId = userId;
+	}
 
-    public String getUsername() {
-        return username;
-    }
+	public String getUsername() {
+		return username;
+	}
 
-    public void setUsername(String username) {
-        this.username = username;
-    }
+	public void setUsername(String username) {
+		this.username = username;
+	}
 
-    public String getFirstName() {
-        return firstName;
-    }
+	public String getFirstName() {
+		return firstName;
+	}
 
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
+	public void setFirstName(String firstName) {
+		this.firstName = firstName;
+	}
 
-    public String getMiddleName() {
-        return middleName;
-    }
+	public String getMiddleName() {
+		return middleName;
+	}
 
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
+	public void setMiddleName(String middleName) {
+		this.middleName = middleName;
+	}
 
-    public String getLastName() {
-        return lastName;
-    }
+	public String getLastName() {
+		return lastName;
+	}
 
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
+	public void setLastName(String lastName) {
+		this.lastName = lastName;
+	}
 
-    public String getPassword() {
-        return password;
-    }
+	public String getPassword() {
+		return password;
+	}
 
-    public void setPassword(String password) {
-        this.password = password;
-    }
+	public void setPassword(String password) {
+		this.password = password;
+	}
 
-    public List<Groups> getGroupsList() {
-        return groupsList;
-    }
+	public List<Groups> getGroupsList() {
+		return groupsList;
+	}
 
-    public void setGroupsList(List<Groups> groupsList) {
-        this.groupsList = groupsList;
-    }
+	public void setGroupsList(List<Groups> groupsList) {
+		this.groupsList = groupsList;
+	}
 
-    public Kunde getKunde() {
-        return kunde;
-    }
+	public Kunde getKunde() {
+		return kunde;
+	}
 
-    public void setKunde(Kunde kundeid) {
-        this.kunde = kundeid;
-    }
+	public void setKunde(Kunde kundeid) {
+		this.kunde = kundeid;
+	}
 
-    @Override
-    public int hashCode() {
-        int hash = 0;
-        hash += (userId != null ? userId.hashCode() : 0);
-        return hash;
-    }
+	@Override
+	public int hashCode() {
+		int hash = 0;
+		hash += (userId != null ? userId.hashCode() : 0);
+		return hash;
+	}
 
-    @Override
-    public boolean equals(Object object) {
-        // TODO: Warning - this method won't work in the case the id fields are not set
-        if (!(object instanceof Users)) {
-            return false;
-        }
-        Users other = (Users) object;
-        if ((this.userId == null && other.userId != null) || (this.userId != null && !this.userId.equals(other.userId))) {
-            return false;
-        }
-        return true;
-    }
+	@Override
+	public boolean equals(Object object) {
+		// TODO: Warning - this method won't work in the case the id fields are not
+		// set
+		if (!(object instanceof Users)) {
+			return false;
+		}
+		Users other = (Users) object;
+		if ((this.userId == null && other.userId != null)
+				|| (this.userId != null && !this.userId.equals(other.userId))) {
+			return false;
+		}
+		return true;
+	}
 
-    @Override
-    public String toString() {
-        return "wbs.corejpa.persistence.Users[ userId=" + userId + " ]";
-    }
-    
+	@Override
+	public String toString() {
+		return "wbs.corejpa.persistence.Users[ userId=" + userId + " ]";
+	}
+
 }
